Start X11 server and dummy sound device on linux
This means that programs that rely on an X11 server (for example
exported projects) will not need to be wrapped with xvfb-run in order to
run.

By default the X11 server will be started on $DISPLAY. If the DISPLAY
environment variable is not set, it will default to ':0' and DISPLAY
will be set.